scripts/requirements.txt
feedparser==6.0.11
psycopg2-binary==2.9.9
pytrends==4.9.2
anthropic==0.25.0
